Initial support for cubemap textures import/export
Fixes #9.
The dds generated on import are correctly read by GIMP and Phothoshop.
There's no sanity checks on the dds texture on export to verify that it
contains all the 6 images plus the indicated mipmaps yet.
If all the dds cubemap and mipmap flags are all set it's expected that
the data contains (<num_mipmaps> * 6) mipmaps.

The (27 * 4) bytes on cubemap textures are left empty since we don't
know yet what they mean. Most of them seem like valid floats.